Aggiunti numerosi controlli per evitare situazione spiacevoli a causa di messaggi non inviati dal server e altre correzioni su errori del ciclo di vita dell'app

+++ Client/app/src/main/java/com/example/cocktailapp/Carrello.java
@@ -78,8 +78,8 @@ public class Carrello {
 
 
     // Calcola il totale del carrello
-    public double calculateTotal() {
-        double total = 0;
+    public float calculateTotal() {
+        float total = 0;
 
         // Calcola il totale dei cocktail
         for (int i = 0; i < beverages.size(); i++) {
